// src/app/models/apoderado.ts
export interface Apoderado {
    nombres: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    email: string;
    clave: string;
    rut: string;
    telefono: number;
	direccion: string;	
	esApoderado: boolean;
    esProfOFunc: boolean;
	esAdmin: boolean;
    cantAlumno: number;
    nombreApSup: string;
    parentescoApSupAlumno: string;
}